lib: move sha256 label to to_fingerprint_string
this brings the separation of work between the fingerprint and
to_fingerprint_string functions for both sha256 and md5 more in line
with one another.

those functions could almost definitely be deduped, but since I don't
think that we will need to add any other hashing functions for
fingerprinting, I think the unfortunate duplication is fine for now.